pyproject.toml
[project]
name = "map-generator-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.7",
    "devtools>=0.12.2",
    "fastapi[standard]>=0.125.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.1",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
